requirements.txt
flask==3.1.1
flask_sqlalchemy==3.1.1
flask_login==0.6.3
flask_wtf==1.2.1
sqlalchemy==2.0.32
python-dotenv==1.0.1
